Do not duplicate logs in -dry-run mode
When -dry-run is specified, all logging configuration is already
replaced by a single configuration printing to console, this
change avoids adding another log Screen appender later.

Also, it changes the printing from stderr to stdout as that's what
is documented to be the destination for the similar -verbose switch.

Closes #59.

Includes testcase.